fix: test failures

+++ frb_example/pure_dart/test/api/newtype_pattern_test.dart
@@ -7,7 +7,7 @@ Future<void> main() async {
 
   test('dart call handleNewtype', () async {
     final newtypeResp = await handleNewtype(arg: NewTypeInt(field0: BigInt.from(42)));
-    expect(newtypeResp.field0, 84);
+    expect(newtypeResp.field0.toInt(), 84);
   });
 
   // TODO rm?
