+++ frb_example/pure_dart/test/api/optional_test.dart
@@ -44,7 +44,7 @@ Future<void> main() async {
       }
       if (ret == null) fail('ret nulled after loop');
       expect(ret.int32, loopFor, reason: 'int32');
-      expect(ret.int64, loopFor, reason: 'int64');
+      expect(ret.int64.toInt(), loopFor, reason: 'int64');
       expect(ret.float64, loopFor, reason: 'float64');
       expect(ret.boolean, false);
       expect(ret.zerocopy?.length, loopFor);
